Switch to new charm microk8s
Update microk8s terraform plan to switch to new
charm microk8s. Add new step to enable addons
in microk8s that should deploy and configure
coredns and metallb charm.
coredns addon is enabled by default in microk8s.
But coredns charm is deployed so that coredns
can be configured in future using charm configs.

Note:
* New charm microk8s deploys only classic snap
but not the strict variant.
* Limitation in charm metallb to scale
https://github.com/charmed-kubernetes/metallb-operator/blob/30038cc35174ffdf944edd3dc18baf6141b5e95c/src/charm.py#L80

TODO:
1. Migration have to be tested from legacy charm
to new charm. To do this, following bugs have
to be resolved
https://github.com/juju/terraform-provider-juju/issues/249
https://github.com/juju/terraform-provider-juju/issues/278

2. Add REAME and tests to migration plugin