Revert "Add manifest support for validation plugin."
This reverts commit 1214b4a52c8bfe3d55638740ee83b55349d4c468.

This caused all the applications to be removed from the openstack model
when running `sunbeam configure validation schedule=<valid schedule>`.
So here we revert this commit, returning to the previous method for
setting the configuration, which was known to be working.

We can implement complete support for manifests in a future patch.