Ensure model is always closed
Each time a model is connected to the controller, it will spawn a background
task with a watcher communicating to the controller. While this can bring many
features, Sunbeam does not make use of them, nor does it want to make use of
them. This can create issues in case of unstable connection, with coroutines
raising exceptions in the background.
Error such as: `Unknown watcher id` or `reconnect error in coro`

Always close a model once usage is done. There are some cases where the
JujuHelper "leak" entities connected to the model. In this case, it's
the caller responsibility to closed the model after usage of these entities is over.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>