Fix cluster refresh
update_tfvars_and_apply_tf removes from tf vars variables that are in
config db but not in manifest, which is fine for most variables as they
are not direct mapping from manifest to tf plan. But it is not the case
for MySQl variables, which are either: computed by sunbeam (not in
manifest) or in manifest, but are used as same name in the config db.
Introduce concept of preserved variable in a tf plan varmap.

Fix feature manager update features instanciation.

Use model from deployment instead of hardcoded names.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>
(cherry picked from commit 2db9ba6c0b5a12b99c4241dae777122c1d45ba57)