Add SSO feature
This change adds SSO enablement in the sunbeam command. There are two
methods through which SSO enablement is done:

* The oauth interface
* The external_provider interface

The oauth interface is used to relate to canonical identity platform or
any other suite that implements the same kind of functionality. The enablement
of this type of integration, requires the consumption of an offer from an
existing model, managed externally. Because we use the remote application
name of these types of integration, to generate a label for this provider,
changes have been made to allow the consumption of a remote offer, with an
alias. Changes have also been made to wait for a SAAS app to be removed
before returning from the existing remove SAAS step. This allows us to
properly add and remove oauth based integrations.

The external_provider is used to integrate IDPs like Google, Okta, Entra
or any other IDP that implements OpenID Connect such as Keycloak. This is
done via an integrator charm that implements the external_provider provider
interface. In this implementation we're reusing the kratos-external-integrator
charm.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>