Add multi-region support
Multi-region deployments aim to provide fault tolerance and reduce
the load on certain components such as databases or AMQP services.

Some of the Openstack services (e.g. Keystone, Horizon) will only
run inside the central region, comprised of nodes that have the
"region-controller" role.

The other nodes will connect to the central region using
cross-controller Juju relations.

The new "region-controller" role is mutually exclusive with all
other roles and can be used to bootstrap a region controller like
so:

```
sunbeam cluster bootstrap --role region-controller
```

Before bootstrapping new regions, a token must be obtained from
the region controller. This registers a new Juju user and provides
the registration token and Juju controller information so that
the other regions may connect to it and establish cross-controller
Juju integrations.

```
sunbeam cluster add-secondary-region-node $nodeName
```

The resulting token can be used to bootstrap the new region like so:

```
sunbeam cluster bootstrap \
  --role control,compute \
  --region-controller-token=$token
```

Note that the Terraform Juju provider does't yet support cross
controller relations, however this feature is expected to be
implemented during this cycle:
https://github.com/juju/terraform-provider-juju/issues/805

In the meantime, the cross-controller relations must be defined
manually. Once the feature gets added to the Terraform provider,
it will be a matter of uncommenting the lines that pass external
Keystone offers and also providing the region controller info.

TODO items, covered in subsequent PRs:

* joining region controller nodes
* joining secondary region nodes
* ensure that certificates are properly handled
* store region information in clusterd and perform additional checks