tls: Add application to list_oustanding_certificates
With manual-tls-certificates 1/stable, the action
get-outstanding-certificates returns list of
CSR and relation_ids. However the same record
(same CSR and relation_id) can be appened to the
list in case of HA for each unit.
relation_map function returns list of units for
each relation_id which is modified to return
application name instead.

In list-outstanding-certificates, the output is
modified to print App, Unit, Relation_id, CSR
with null values if information is not available.
Unit is not entirely removed for backward
compatibility.
For the prompt questions in sunbeam tls ca unit_certs,
the prompt question ask for certificate for the
app instead of unit. For backward compatiblity,
if the unit name is available, it is used in
prompt question.
Currently, the traefik and vault requests for App
level certificate to manual-tls-certificates and
so we might not end up with unit names with
manual-tls-certificates 1/stable (to be precise
tls-certificates-interface v4 library). However
unit_name is supported for backward compatiblity.

Note the output for list_outstanding_certificates:

- app_name: traefik
  unit_name: null
  relation_id: 123
  csr: <CSR>