Avoid terraform schema changes
Some services will be disabled on multi-region environments.
For example, Keystone and Horizon are the only services expected
to run on region controller nodes.

Normally we'd just update the Terraform plans and exclude these
services (e.g. using the "count" Terraform meta-argument).
Unfortunately that would cause a Terraform schema change, transforming
modules into tuples and causing upgrade issues.

Sunbeam cannot handle Terraform schema changes, as such we have to
completely avoid these. In this case, we'll always deploy the
disabled charms, but set the scale to 0.

This commit updates the Sunbeam steps to ignore applications that
are expected to remain in blocked state.