instance-recovery: Add storage IP pool support for maas deployment
consul-k8s charm is updated to create a loadbalancer for
gossip ports and send the information to consul-client
via consul-cluster relation data. The loadbalancer can
be created based on configuration option in consul-k8s
charm.

sunbeam can enable the option only for the consul-storage
application (which will come into picture only when the
storage space is separate). This option is not required
for management space as the current endpoints provided
to consul client are from management space.
Users have to define a reserved pool (1 IP is good) in
storage space with comment <cloud>-storage-ippool.
sunbeam creates corresponding metallb ipaddress pool.
Once the loadbalancer service is created for
consul-storage with name consul-storage-lb, sunbeam
will update the metallb annotations to use the
storage pool for loadbalancer IP.

Modifications:
* Add storage ip pool label `<cloud>-storage-ippool` in
MaasDeployment
* Update Creation of IPAddressPools in maas to include
storage ip pool if exists. Do not error the step if the
storage ip pool does not exist in maas.
* Update instance-recovery feature to rename consul-client
config enable-tcp-health-check to enable-health-check
* Update PatchLoadBalancerIPStep, PatchLoadBalancerIPStep
to include new parameter ignore_errors that will ignore
if the service does not exist and not of type loadbalancer.
* Add PatchLoadbalcerIP/IPPool Step to feature instance-recovery
and add them to plan only if the deployment is of type maas.
* Use kubectl apply instead of patch in PatchLoadbalcerIP/IPPool
Step as patch does not work with multiple ports with same target
even though protocol is different.
https://github.com/kubernetes/kubernetes/issues/105610