feat(storage): add gate flag on storage backends
Add a storage flag gate for experimental backends.
A backend can be enabled by using:
`sudo snap set openstack storage.<backend>=true`

Once the backend is deployed, it will be added to clusterd as an enabled
backends, and will be accessible on other instances of the openstack
cli, and used to determine if the backend is available.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>